+++ lib/flutter_map_tile_caching.dart
@@ -1,24 +1,25 @@
 // This is flutter_map_tile_caching's main file
 // You should never need to include anything other than this file from this library in your project
 
-// Regions
-export 'src/regions/circle.dart';
-export 'src/regions/downloadableRegion.dart';
-export 'src/regions/line.dart';
-export 'src/regions/rectangle.dart';
-
 // Main
 export 'src/main.dart';
 //export 'src/tileLayer.dart'; EXPERIMENTAL
 
-// Backend and Extras
+// Backend and Misc
 export 'src/storageManager.dart';
-export 'src/exts.dart';
+export 'src/misc.dart';
+
+// Regions
+export 'src/regions/downloadableRegion.dart';
+export 'src/regions/recoveredRegion.dart';
+export 'src/regions/rectangle.dart';
+export 'src/regions/circle.dart';
+export 'src/regions/line.dart';
 
 // Widgets
-export 'src/widgets/shapeChooser.dart';
 //export 'src/widgets/cacheScreen.dart'; EXPERIMENTAL
 //export 'src/widgets/sourceSwitcher.dart'; EXPERIMENTAL
 
 // Deprecated
 export 'src/deprecated/oldCachingManager.dart';
+export 'src/deprecated/shapeChooser.dart';